Add new runner module
