+++ src/vivarium_cluster_tools/dagger/runner.py
@@ -1,0 +1,113 @@
+"""
+=============
+dagger Runner
+=============
+
+The main process loop for ``dagger run`` invocations: parse the
+workflow configuration, build the Jobmon workflow, bind and run it,
+and notify on completion.
+"""
+
+from __future__ import annotations
+
+import hashlib
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any
+
+import yaml
+from loguru import logger
+
+from vivarium_cluster_tools.notifications import send_slack_notification
+from vivarium_cluster_tools.psimulate.jobmon_config import client
+from vivarium_cluster_tools.psimulate.workflow_config.builder import (
+    build_workflow_from_config,
+)
+from vivarium_cluster_tools.psimulate.workflow_config.config import WorkflowConfig
+from vivarium_cluster_tools.psimulate.workflow_config.serialization import (
+    workflow_config_to_dict,
+)
+from vivarium_cluster_tools.psimulate.workflow_config.utilities import WORKFLOW_ARGS_FILENAME
+
+
+def workflow_main(
+    workflow_config: WorkflowConfig,
+    verbose: int = 0,
+    resume: bool = False,
+) -> None:
+    """Entry point for the ``dagger run`` subcommand.
+
+    Parameters
+    ----------
+    workflow_config
+        The parsed and validated workflow configuration (with CLI overrides applied).
+    verbose
+        Verbosity level.
+    resume
+        Whether to resume a previously started workflow.
+    """
+    logger.info(f"Starting workflow: {workflow_config.name}")
+
+    # Create output directory if it doesn't exist
+    output_root = workflow_config.output_directory
+    output_root.mkdir(parents=True, exist_ok=True)
+
+    workflow_args_path = output_root / WORKFLOW_ARGS_FILENAME
+
+    if resume:
+        workflow_args = workflow_args_path.read_text().strip()
+        logger.info(f"Resuming workflow with args: {workflow_args}")
+    else:
+        # Generate a unique workflow_args using a timestamp so each fresh
+        # run is distinct even with the same config and output directory.
+        timestamp = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
+        output_hash = hashlib.md5(str(output_root).encode()).hexdigest()[:8]
+        workflow_args = f"workflow_{workflow_config.name}_{output_hash}_{timestamp}"
+
+    # Write the requested configuration to output directory
+    write_workflow_configuration(output_root, workflow_config)
+
+    # Build the workflow
+    logger.debug("Building workflow.")
+    workflow = build_workflow_from_config(workflow_config, workflow_args=workflow_args)
+
+    # Persist workflow_args before running so --resume can find it
+    workflow_args_path.write_text(workflow_args)
+
+    wf_status, monitoring_url = client.bind_and_run_workflow(
+        workflow, output_root, resume=resume
+    )
+
+    send_slack_notification(
+        workflow_name=workflow_config.name,
+        status=wf_status,
+        command_label="dagger run",
+        monitoring_url=monitoring_url,
+        results_dir=str(output_root),
+    )
+
+    if wf_status != client.JOBMON_STATUS_DONE:
+        raise RuntimeError(
+            f"Workflow finished with status '{wf_status}' "
+            f"(expected '{client.JOBMON_STATUS_DONE}' for DONE)."
+        )
+    logger.info(f"Workflow completed successfully. Results in {output_root}")
+
+
+def write_workflow_configuration(output_root: Path, workflow_config: WorkflowConfig) -> None:
+    """Write workflow configuration to a YAML file in the output directory.
+
+    Creates a ``configuration.yaml`` that can be reused directly with
+    ``dagger run --config configuration.yaml``.
+
+    Parameters
+    ----------
+    output_root
+        The root output directory for the workflow.
+    workflow_config
+        The parsed and validated workflow configuration.
+    """
+    config: dict[str, Any] = {"workflow": workflow_config_to_dict(workflow_config)}
+    config_file = output_root / "configuration.yaml"
+    config_file.write_text(yaml.dump(config, default_flow_style=False, sort_keys=False))
+    logger.info(f"Run configuration written to {config_file}")
