+++ src/vivarium_cluster_tools/psimulate/runner.py
@@ -12,7 +12,6 @@ from __future__ import annotations
 import hashlib
 import os
 import shutil
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
 
@@ -39,76 +38,7 @@ from vivarium_cluster_tools.psimulate.performance_logger import (
     append_perf_data_to_central_logs,
 )
 from vivarium_cluster_tools.psimulate.results.writing import collect_metadata
-from vivarium_cluster_tools.psimulate.workflow_config.builder import (
-    build_workflow_from_config,
-)
-from vivarium_cluster_tools.psimulate.workflow_config.config import WorkflowConfig
-from vivarium_cluster_tools.psimulate.workflow_config.serialization import (
-    workflow_config_to_dict,
-)
-from vivarium_cluster_tools.psimulate.workflow_config.utilities import WORKFLOW_ARGS_FILENAME
 from vivarium_cluster_tools.vipin.perf_report import report_performance
-
-def workflow_main(
-    workflow_config: WorkflowConfig,
-    verbose: int = 0,
-    resume: bool = False,
-) -> None:
-    """Entry point for the psimulate workflow subcommand.
-
-    Parameters
-    ----------
-    workflow_config
-        The parsed and validated workflow configuration (with CLI overrides applied).
-    verbose
-        Verbosity level.
-    resume
-        Whether to resume a previously started workflow.
-    """
-    logger.info(f"Starting workflow: {workflow_config.name}")
-
-    # Create output directory if it doesn't exist
-    output_root = workflow_config.output_directory
-    output_root.mkdir(parents=True, exist_ok=True)
-
-    workflow_args_path = output_root / WORKFLOW_ARGS_FILENAME
-
-    if resume:
-        workflow_args = workflow_args_path.read_text().strip()
-        logger.info(f"Resuming workflow with args: {workflow_args}")
-    else:
-        # Generate a unique workflow_args using a timestamp so each fresh
-        # run is distinct even with the same config and output directory.
-        timestamp = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
-        output_hash = hashlib.md5(str(output_root).encode()).hexdigest()[:8]
-        workflow_args = f"workflow_{workflow_config.name}_{output_hash}_{timestamp}"
-
-    # Write the requested configuration to output directory
-    write_workflow_configuration(output_root, workflow_config)
-
-    # Build the workflow
-    logger.debug("Building workflow.")
-    workflow = build_workflow_from_config(workflow_config, workflow_args=workflow_args)
-
-    # Persist workflow_args before running so --resume can find it
-    workflow_args_path.write_text(workflow_args)
-
-    wf_status, monitoring_url = client.bind_and_run_workflow(workflow, output_root, resume=resume)
-
-    send_slack_notification(
-        workflow_name=workflow_config.name,
-        status=wf_status,
-        command_label="psimulate workflow",
-        monitoring_url=monitoring_url,
-        results_dir=str(output_root),
-    )
-
-    if wf_status != client.JOBMON_STATUS_DONE:
-        raise RuntimeError(
-            f"Workflow finished with status '{wf_status}' "
-            f"(expected '{client.JOBMON_STATUS_DONE}' for DONE)."
-        )
-    logger.info(f"Workflow completed successfully. Results in {output_root}")
 
 
 def report_initial_status(
@@ -169,25 +99,6 @@ def write_backup_metadata(
         mode="a",
         header=not os.path.exists(backup_metadata_path),
     )
-
-
-def write_workflow_configuration(output_root: Path, workflow_config: WorkflowConfig) -> None:
-    """Write workflow configuration to a YAML file in the output directory.
-
-    Creates a ``configuration.yaml`` that can be reused directly with
-    ``psimulate workflow -c configuration.yaml``.
-
-    Parameters
-    ----------
-    output_root
-        The root output directory for the workflow.
-    workflow_config
-        The parsed and validated workflow configuration.
-    """
-    config: dict[str, Any] = {"workflow": workflow_config_to_dict(workflow_config)}
-    config_file = output_root / "configuration.yaml"
-    config_file.write_text(yaml.dump(config, default_flow_style=False, sort_keys=False))
-    logger.info(f"Run configuration written to {config_file}")
 
 
 def write_configuration(
