+++ tests/psimulate/test_cli.py
@@ -616,203 +616,3 @@ class TestDeprecationDeadline:
             f"--model-specification/-M, --branch-configuration/-B, and "
             f"--results-root/-R exclusively."
         )
-
-
-class TestWorkflowSubcommand:
-    """Tests for -c/--config on the ``workflow`` subcommand."""
-
-    def test_workflow_with_config_file(self, tmp_path: Path) -> None:
-        """Workflow subcommand accepts -c flag with workflow config."""
-        workflow_yaml = _write_yaml(
-            tmp_path,
-            {
-                "workflow": {
-                    "name": "test_workflow",
-                    "project": "proj_simscience",
-                    "queue": "all.q",
-                    "output_directory": str(tmp_path / "output"),
-                    "steps": [
-                        {
-                            "name": "test_step",
-                            "command": "pytest tests/",
-                            "resources": {"memory_gb": 4, "runtime": "01:00:00"},
-                        }
-                    ],
-                }
-            },
-            name="pipeline.yaml",
-        )
-
-        cli_runner = CliRunner()
-        with patch(
-            "vivarium_cluster_tools.psimulate.runner.workflow_main"
-        ) as mock_workflow_main:
-            result = cli_runner.invoke(psimulate, ["workflow", "-c", str(workflow_yaml)])
-
-        assert result.exit_code == 0, result.output
-        mock_workflow_main.assert_called_once()
-        call_kwargs = mock_workflow_main.call_args.kwargs
-        assert call_kwargs["workflow_config"].name == "test_workflow"
-        assert call_kwargs["workflow_config"].project == "proj_simscience"
-
-    @pytest.mark.parametrize(
-        "cli_args,expected_overrides",
-        [
-            # Single flag overrides
-            (["-P", "proj_simscience_prod"], {"project": "proj_simscience_prod"}),
-            (["--project", "proj_simscience_prod"], {"project": "proj_simscience_prod"}),
-            (["-q", "long.q"], {"queue": "long.q"}),
-            (["--queue", "long.q"], {"queue": "long.q"}),
-            # Multiple flags overridden simultaneously
-            (
-                ["-P", "proj_simscience_prod", "-q", "long.q"],
-                {"project": "proj_simscience_prod", "queue": "long.q"},
-            ),
-        ],
-    )
-    def test_workflow_cli_overrides_config(
-        self,
-        tmp_path: Path,
-        cli_args: list[str],
-        expected_overrides: dict[str, str],
-    ) -> None:
-        """CLI flags override corresponding values from config file.
-
-        Tests both individual flag overrides and multiple simultaneous overrides.
-        """
-        workflow_yaml = _write_yaml(
-            tmp_path,
-            {
-                "workflow": {
-                    "name": "test_workflow",
-                    "project": "proj_simscience",
-                    "queue": "all.q",
-                    "output_directory": str(tmp_path / "output"),
-                    "steps": [
-                        {"name": "s1", "command": "echo test", "resources": {"memory_gb": 2}}
-                    ],
-                }
-            },
-            name="pipeline.yaml",
-        )
-
-        cli_runner = CliRunner()
-        with patch(
-            "vivarium_cluster_tools.psimulate.runner.workflow_main"
-        ) as mock_workflow_main:
-            result = cli_runner.invoke(
-                psimulate,
-                ["workflow", "-c", str(workflow_yaml)] + cli_args,
-            )
-
-        assert result.exit_code == 0, result.output
-        call_kwargs = mock_workflow_main.call_args.kwargs
-
-        # Verify all expected overrides were applied
-        for config_key, expected_value in expected_overrides.items():
-            assert getattr(call_kwargs["workflow_config"], config_key) == expected_value
-
-    def test_workflow_cli_overrides_output_directory(self, tmp_path: Path) -> None:
-        """CLI -o flag overrides output_directory from config file."""
-        original_output = tmp_path / "output_original"
-        override_output = tmp_path / "output_override"
-        override_output.mkdir()
-
-        workflow_yaml = _write_yaml(
-            tmp_path,
-            {
-                "workflow": {
-                    "name": "test_workflow",
-                    "project": "proj_simscience",
-                    "queue": "all.q",
-                    "output_directory": str(original_output),
-                    "steps": [
-                        {"name": "s1", "command": "echo test", "resources": {"memory_gb": 2}}
-                    ],
-                }
-            },
-            name="pipeline.yaml",
-        )
-
-        cli_runner = CliRunner()
-        with patch(
-            "vivarium_cluster_tools.psimulate.runner.workflow_main"
-        ) as mock_workflow_main:
-            result = cli_runner.invoke(
-                psimulate,
-                ["workflow", "-c", str(workflow_yaml), "-o", str(override_output)],
-            )
-
-        assert result.exit_code == 0, result.output
-        call_kwargs = mock_workflow_main.call_args.kwargs
-        assert call_kwargs["workflow_config"].output_directory == override_output.resolve()
-
-    def test_workflow_missing_config_flag_errors(self) -> None:
-        """Running workflow without -c flag produces an error.
-
-        The -c/--config flag is required for the workflow subcommand.
-        Click should raise an error if it's missing.
-        """
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(psimulate, ["workflow"])
-        assert result.exit_code != 0
-        # Click's error message for missing required options typically mentions "Missing option"
-        assert "missing" in result.output.lower() or "required" in result.output.lower()
-
-    def test_workflow_project_required(self, tmp_path: Path) -> None:
-        """Project is required; error if missing from both config and CLI."""
-        workflow_yaml = _write_yaml(
-            tmp_path,
-            {
-                "workflow": {
-                    "name": "test_workflow",
-                    # Note: project is NOT provided
-                    "queue": "all.q",
-                    "output_directory": str(tmp_path / "output"),
-                    "steps": [
-                        {
-                            "name": "test_step",
-                            "command": "echo test",
-                            "resources": {"memory_gb": 2},
-                        }
-                    ],
-                }
-            },
-            name="pipeline.yaml",
-        )
-
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(psimulate, ["workflow", "-c", str(workflow_yaml)])
-        assert result.exit_code != 0
-        error_message = str(result.exception).lower()
-        assert "project" in error_message
-        assert "required" in error_message
-
-    def test_workflow_output_directory_required(self, tmp_path: Path) -> None:
-        """Output directory is required; error if missing from both config and CLI."""
-        workflow_yaml = _write_yaml(
-            tmp_path,
-            {
-                "workflow": {
-                    "name": "test_workflow",
-                    "project": "proj_simscience",
-                    "queue": "all.q",
-                    # NOTE: output_directory is NOT provided
-                    "steps": [
-                        {
-                            "name": "test_step",
-                            "command": "echo test",
-                            "resources": {"memory_gb": 2},
-                        }
-                    ],
-                }
-            },
-            name="pipeline.yaml",
-        )
-
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(psimulate, ["workflow", "-c", str(workflow_yaml)])
-        assert result.exit_code != 0
-        error_message = str(result.exception).lower()
-        assert "output" in error_message
-        assert "required" in error_message
